fix: generating inclusion proofs with same params for verifying proof on Solidity verifier
