+++ contracts/test/Verifiers.ts
@@ -9,12 +9,6 @@ import * as fs from "fs";
 import * as path from "path";
 
 describe("Verifier Contracts", () => {
-  async function deployBN256G2() {
-    const bn256g2 = await ethers.getContractFactory("BN256G2");
-    const bn256g2Instance = await bn256g2.deploy();
-    return bn256g2Instance;
-  }
-
   async function deployVerifyingFixture() {
     // Contracts are deployed using the first signer/account by default
     const verifyingKey = await ethers.deployContract(
@@ -30,40 +24,6 @@ describe("Verifier Contracts", () => {
     };
   }
 
-  describe("BN256G2 operation test", () => {
-    it("should test addition", async () => {
-      const bn256g2 = await deployBN256G2();
-      let res = await bn256g2.ECTwistAdd(BigNumber.from("0x010acd800ef6f6752c7fd63efd27310fbbc0cddeafd19df13772b0162dd60198"),
-        BigNumber.from("0x11fb65fc8359f0a2fbc2076b919b10cd6a0259be9b982c77749f6612bd84f619"),
-        BigNumber.from("0x1842bf644523eb1b63e72b333290bdd8e71b61217d423d03e76035f5672dca3b"),
-        BigNumber.from("0x248e3efe73cd2286e12e6649a4e3a6557d6ddb21dab6cdeb6af71e83e54f5c39"),
-        BigNumber.from("0x0d76cdeac516681d6ee49f97ea2449a332da02f2932820e780672d858c0b9e67"),
-        BigNumber.from("0x23cf199ce510a7bd88d1a36d01d27a5d20877a089257cee26e9e3da130e08a56"),
-        BigNumber.from("0x1e2d1cb11f0c832929d64e7e2d9497344a5756f5ee4144186b7384c5babeb2d4"),
-        BigNumber.from("0x2e01ba57b5e27d864f0b161e9fb642373bc2a0cb009114a504fe87ce46d2362e"));
-
-      expect(res[0]).to.be.equal(BigNumber.from("0x123aeb5d388385c95f621887198c49ce360be39e202f1e5b2cb716fb16a2947e"));
-      expect(res[1]).to.be.equal(BigNumber.from("0x187aabbe9dbc6b188789e8063c6848253365b628774cadfa2f5499c389bc720d"));
-      expect(res[2]).to.be.equal(BigNumber.from("0xe1d1cc06f98a2603ed2b548b4776241a444b38bc3066023cdb27fe567849116"));
-      expect(res[3]).to.be.equal(BigNumber.from("0x1671331a9c903755d2517f0f98e8fbf84b025ab5e50b89866974556ff86827f2"));
-    });
-    it("should test multiplication", async () => {
-      const bn256g2 = await deployBN256G2();
-
-      let res = await bn256g2.ECTwistMul(
-        BigNumber.from("0x08b88ebb6b6820ebe287214692ad2b2aed666a4b0c6375a73a63c16264d1bf64"),
-        BigNumber.from("0x1800deef121f1e76426a00665e5c4479674322d4f75edadd46debd5cd992f6ed"),
-        BigNumber.from("0x198e9393920d483a7260bfb731fb5d25f1aa493335a9e71297e485b7aef312c2"),
-        BigNumber.from("0x12c85ea5db8c6deb4aab71808dcb408fe3d1e7690c43d37b4ce6cc0166fa7daa"),
-        BigNumber.from("0x090689d0585ff075ec9e99ad690c3395bc4b313370b38ef355acdadcd122975b"));
-
-      expect(res[0]).to.be.equal(BigNumber.from("0x010acd800ef6f6752c7fd63efd27310fbbc0cddeafd19df13772b0162dd60198"));
-      expect(res[1]).to.be.equal(BigNumber.from("0x11fb65fc8359f0a2fbc2076b919b10cd6a0259be9b982c77749f6612bd84f619"));
-      expect(res[2]).to.be.equal(BigNumber.from("0x1842bf644523eb1b63e72b333290bdd8e71b61217d423d03e76035f5672dca3b"));
-      expect(res[3]).to.be.equal(BigNumber.from("0x248e3efe73cd2286e12e6649a4e3a6557d6ddb21dab6cdeb6af71e83e54f5c39"));
-    });
-
-  });
 
   describe("Snark Proof Verifier", () => {
     let snarkVerifier: SnarkVerifier;
@@ -71,6 +31,7 @@ describe("Verifier Contracts", () => {
     let commitmentCalldata: {
       range_check_snark_proof: BytesLike;
       grand_sums_batch_proof: BytesLike;
+      total_balances: BigNumber[];
     };
     beforeEach(async () => {
       const deploymentInfo = await loadFixture(deployVerifyingFixture);
@@ -90,7 +51,7 @@ describe("Verifier Contracts", () => {
       expect(await snarkVerifier.verifyProof(verifyingKey.address, commitmentCalldata.range_check_snark_proof, [1])).to.be.true;
     });
 
-    it("hould fail to verify snark proof without the number of instances", async () => {
+    it("should fail to verify snark proof without the number of instances", async () => {
       await expect(snarkVerifier.verifyProof(verifyingKey.address, commitmentCalldata.range_check_snark_proof, [])).to.be.reverted;
     });
 
@@ -102,6 +63,7 @@ describe("Verifier Contracts", () => {
     let commitmentCalldata: {
       range_check_snark_proof: BytesLike;
       grand_sums_batch_proof: BytesLike;
+      total_balances: BigNumber[];
     };
 
     beforeEach(async () => {
@@ -119,6 +81,7 @@ describe("Verifier Contracts", () => {
       // Concatenates the snark proof and the grand sum proof
       let snarkProofArray = ethers.utils.arrayify(commitmentCalldata.range_check_snark_proof);
       let grandSumProofArray = ethers.utils.arrayify(commitmentCalldata.grand_sums_batch_proof);
+      let totalBalances = commitmentCalldata.total_balances;
 
       // The first 64 bytes of the snark proof represent a commitment to the corresponding username polynomial
       // Starting from the next 64 bytes, each set of 64 bytes represents commitments corresponding to the total sum of balances
@@ -132,7 +95,7 @@ describe("Verifier Contracts", () => {
       //  Where `N` is the number of currencies
       let proofs = ethers.utils.hexlify(ethers.utils.concat([grandSumProofArray, grandSumCommitments]));
 
-      expect(await grandSumVerifier.verifyProof(verifyingKey.address, proofs, [])).to.be.true;
+      expect(await grandSumVerifier.verifyProof(verifyingKey.address, proofs, totalBalances)).to.be.true;
     });
   });
 
@@ -144,8 +107,9 @@ describe("Verifier Contracts", () => {
     let commitmentCalldata: {
       range_check_snark_proof: BytesLike;
       grand_sums_batch_proof: BytesLike;
+      total_balances: BigNumber[];
     };
-    let challenge: BytesLike;
+    let challenges: [BigNumber, BigNumber, BigNumber, BigNumber];
     let username: BytesLike;
     let username_biguint: BigNumber;
     let balance1: BigNumber;
@@ -157,9 +121,9 @@ describe("Verifier Contracts", () => {
       commitmentCalldata = deploymentInfo.commitmentCalldata;
 
       // InclusionVerifier requires BN256G2 contract for performing elliptic curve operations on G2 subgroup
-      const bn256g2 = await deployBN256G2();
+      // const bn256g2 = await deployBN256G2();
       inclusionVerifier = await ethers.deployContract(
-        "src/InclusionVerifier.sol:InclusionVerifier", [bn256g2.address]
+        "src/InclusionVerifier.sol:InclusionVerifier"
       ) as InclusionVerifier;
       await inclusionVerifier.deployed();
 
@@ -176,14 +140,13 @@ describe("Verifier Contracts", () => {
 
       inclusionProof = inclusionCalldata.proof;
       username = inclusionCalldata.username;
-      challenge = inclusionCalldata.challenge;
-      username_biguint = inclusionCalldata.balances[0];
-      balance1 = inclusionCalldata.balances[1];
-      balance2 = inclusionCalldata.balances[2];
+      challenges = inclusionCalldata.challenges;
+      username_biguint = inclusionCalldata.user_values[0];
+      balance1 = inclusionCalldata.user_values[1];
+      balance2 = inclusionCalldata.user_values[2];
     });
 
     it("should verify inclusion proof", async () => {
-
       // Generating proof with concatenated snark proof and inclusion proof
       let snarkProof = commitmentCalldata.range_check_snark_proof;
 
@@ -196,8 +159,8 @@ describe("Verifier Contracts", () => {
 
       expect(await inclusionVerifier.verifyProof(
         verifyingKey.address,
-        challenge,
         proofs,
+        [challenges[0], challenges[1], challenges[2], challenges[3]],
         [username_biguint, balance1, balance2]
       )).to.be.true;
     });
