+++ kzg_prover/bin/inclusion_proof_solidity_calldata.json
@@ -1,10 +1,15 @@
 {
-  "proof": "0x05005ce04db0984cb154ec9c7f605b352c283273bb5032e24b7aa0afc95ebc5608436cab900bc8e5ac406298b5076ea8d2fd4cafe36b71d0c72ec71edcf0886823293705b1b7ae6c642f92c33594fd57900d5f5bfab486e1da4ec55c9ddc8f070434950bbdec3e404e8a3f63166d6271a98b6b49b9927482d2e197183b046214100b83bdf0b745234eed66918ac1e91d8207152104d4d0fa57f917032cc018402c5c817b515d5214b772f02fde0d8b407881eaddba64c9b8259145d9a0eda16a",
-  "challenge": "0x2691822789aea440946baae8421ce219823e0f4cf0b0545689249d72bca7b4d8",
-  "username": "lAhWlEWZ",
-  "balances": [
-    "0x6c4168576c45575a",
-    "0x48db",
-    "0x827"
+  "proof": "0x1e7863f0a14846a0fb8b20db94ee4d6fc75a30c6942bcf3d1ca70a80ad3ac5ae21f43295f233ff1f6cc47f55795a209f9ca734c76c2d2c3bcb37d955b58702ca0938c672375247acf7a110224e1f3d0e118b289a02c8b4a0acd33a8d209dbebc01895e90be59da66f1de8e8a144047936e7393e312fbbcfe11960be9de661277150f79ec188c492fc71556342c001c2cdf89aca2df24b2f68514e1742185c2021f0479d9e8e1a3e1c14e2bac6af1c5ebd3d6c5b35c230c0cc92c554cc245a9d5",
+  "challenges": [
+    "0xf79a0045992596e3278606b5317aaf4f6bb65071219b1c89d542509fe6dddd3",
+    "0x2299faaf0e21893e99005dc9165fba869b5aa88bcac5af4395071fd569686fde",
+    "0x2480ccf8a3834d03badc6b0c35773d758908019658fe9ea3bf85dc202257cfda",
+    "0x2bd763bd08804bb437e62575eff584497632e3d3034a20f2a868b1136e4adb6e"
+  ],
+  "username": "MBlfbBGI",
+  "user_values": [
+    "0x4d426c6662424749",
+    "0x108ef",
+    "0x48db"
   ]
 }
